+++ apps/angular-test-app/src/app/app.component.ts
@@ -1,12 +1,18 @@
 import { Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { AngularLibComponent } from '@angular-java-monorepo/angular-lib';
+import { AngularLibComponent } from '@gipo999/angular-lib';
+import { AngularLibThreeComponent } from '@gipo999/angular-lib-three';
 
 import { NxWelcomeComponent } from './nx-welcome.component';
 
 @Component({
   standalone: true,
-  imports: [NxWelcomeComponent, RouterModule, AngularLibComponent],
+  imports: [
+    NxWelcomeComponent,
+    RouterModule,
+    AngularLibComponent,
+    AngularLibThreeComponent,
+  ],
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',