+++ libs/angular-lib-three/lint-staged.config.js
@@ -1,0 +1,4 @@
+/* eslint-disable unicorn/prefer-module */
+const defaultBase = require('../../lint-staged.config');
+
+module.exports = defaultBase;