+++ libs/angular-lib-three/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/angular-lib-three/angular-lib-three.component';