+++ apps/angular-test-app/CHANGELOG.md
@@ -1,29 +1,49 @@
 ## 1.1.0-next.3 (2024-07-03)
 
-This was a version bump only for angular-test-app to align it with other projects, there were no code changes.
+## 1.1.4
+
+## 1.1.4-next.0
+
+## 1.1.3
+
+## 1.1.2
+
+## 1.1.1
+
+### Patch Changes
+
+- a6fbc62: manually fixed versions
+- 634a27b: first versioning
+
+This was a version bump only for angular-test-app to align it with other
+projects, there were no code changes.
 
 ## 1.1.0-next.2 (2024-07-03)
 
-This was a version bump only for angular-test-app to align it with other projects, there were no code changes.
+This was a version bump only for angular-test-app to align it with other
+projects, there were no code changes.
 
 ## 1.1.0-next.1 (2024-07-03)
 
-This was a version bump only for angular-test-app to align it with other projects, there were no code changes.
+This was a version bump only for angular-test-app to align it with other
+projects, there were no code changes.
 
 ## 1.1.0-next.0 (2024-07-03)
 
-This was a version bump only for angular-test-app to align it with other projects, there were no code changes.
+This was a version bump only for angular-test-app to align it with other
+projects, there were no code changes.
 
 # 1.0.0 (2024-07-03)
 
-This was a version bump only for angular-test-app to align it with other projects, there were no code changes.
+This was a version bump only for angular-test-app to align it with other
+projects, there were no code changes.
 
 ## 0.1.0 (2024-07-03)
 
-This was a version bump only for angular-test-app to align it with other projects, there were no code changes.
+This was a version bump only for angular-test-app to align it with other
+projects, there were no code changes.
 
 ## 0.0.1 (2024-07-03)
-
 
 ### 🚀 Features
 
@@ -39,8 +59,7 @@ This was a version bump only for angular-test-app to align it with other project
 
 - nx release with changelog and github release
 
-
-### ❤️  Thank You
+### ❤️ Thank You
 
 - Gipo
 - gipo355